requirements.txt
yfinance
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
nltk
requests
python-dotenv
